requirements.txt
aiofiles==25.1.0
flask
pillow
playwright==1.57.0